좋아요 및 일부 에러코드 추가

+++ MSA/Recruitment/src/main/java/com/springcooler/sgma/recruitmentboardlike/command/application/service/RecruitmentBoardLikeServiceImpl.java
@@ -1,5 +1,7 @@
 package com.springcooler.sgma.recruitmentboardlike.command.application.service;
 
+import com.springcooler.sgma.recruitmentboard.command.domain.aggregate.RecruitmentBoard;
+import com.springcooler.sgma.recruitmentboard.command.domain.repository.RecruitmentBoardRepository;
 import com.springcooler.sgma.recruitmentboardlike.command.domain.aggregate.RecruitmentBoardLike;
 import com.springcooler.sgma.recruitmentboardlike.command.domain.repository.RecruitmentBoardLikeRepository;
 import com.springcooler.sgma.recruitmentboardlike.common.exception.CommonException;
@@ -12,6 +14,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 
 
 @Service
@@ -20,11 +23,13 @@ import java.util.List;
 public class RecruitmentBoardLikeServiceImpl implements RecruitmentBoardLikeCommandService {
     private final RecruitmentBoardLikeService recruitmentBoardLikeService;
     private final RecruitmentBoardLikeRepository recruitmentBoardLikeRepository;
+    private final RecruitmentBoardRepository recruitmentBoardRepository;
 
 
     @Transactional
     @Override
     public RecruitmentBoardLike addLike(Long recruitmentBoardId, Long userId) {
+        Optional<RecruitmentBoard> optionalRecruitmentBoard = recruitmentBoardRepository.findById(recruitmentBoardId);
         List<RecruitmentBoardLikeDTO> recruitmentBoardLikes = recruitmentBoardLikeService.findAllRecruitmentBoardLike();
         for (RecruitmentBoardLikeDTO like : recruitmentBoardLikes) {
             if (like.getRecruitmentBoardId().equals(recruitmentBoardId) && like.getUserId().equals(userId)) {
@@ -34,28 +39,42 @@ public class RecruitmentBoardLikeServiceImpl implements RecruitmentBoardLikeComm
                     }
                 }
             }
-            RecruitmentBoardLike recruitmentLike = RecruitmentBoardLike.builder()
-                    .recruitmentBoardId(recruitmentBoardId)
-                    .userId(userId)
-                    .build();
-            System.out.println("좋아요 추가 성공!");
-        return recruitmentBoardLikeRepository.save(recruitmentLike);
+        RecruitmentBoard recruitmentBoard = optionalRecruitmentBoard.get();
+
+        RecruitmentBoard updatedRecruitmentBoard = recruitmentBoard.toBuilder()
+                .likes(recruitmentBoard.getLikes()+1)
+                .build();
+        RecruitmentBoardLike recruitmentLike = RecruitmentBoardLike.builder()
+                .recruitmentBoardId(recruitmentBoardId)
+                .userId(userId)
+                .build();
+        recruitmentBoardRepository.save(updatedRecruitmentBoard);
+        recruitmentBoardLikeRepository.save(recruitmentLike);
+        return null;
     }
 
     @Transactional
     @Override
     public RecruitmentBoardLike deleteLike(Long recruitmentBoardId, Long userId) {
+        Optional<RecruitmentBoard> optionalRecruitmentBoard = recruitmentBoardRepository.findById(recruitmentBoardId);
         List<RecruitmentBoardLikeDTO> recruitmentBoardLikes = recruitmentBoardLikeService.findAllRecruitmentBoardLike();
-
-        for (RecruitmentBoardLikeDTO like : recruitmentBoardLikes) {
-            if (like.getRecruitmentBoardId().equals(recruitmentBoardId) && like.getUserId().equals(userId)) {
-                RecruitmentBoardLike recruitmentLike = recruitmentBoardLikeRepository.findByRecruitmentBoardIdAndUserId(recruitmentBoardId, userId);
-                recruitmentBoardLikeRepository.delete(recruitmentLike);
-                System.out.println("좋아요가 취소되었습니다.");
-                return null;
-            }
-            else{
-                throw new CommonException(ErrorCode.NOT_FOUND_LIKE);
+        if(recruitmentBoardLikes.size()==0) {
+            throw new CommonException(ErrorCode.NOT_FOUND_LIKE);
+        }
+        else {
+            for (RecruitmentBoardLikeDTO like : recruitmentBoardLikes) {
+                if (like.getRecruitmentBoardId().equals(recruitmentBoardId) && like.getUserId().equals(userId)) {
+                    RecruitmentBoard recruitmentBoard = optionalRecruitmentBoard.get();
+                    RecruitmentBoard updatedRecruitmentBoard = recruitmentBoard.toBuilder()
+                            .likes(recruitmentBoard.getLikes() - 1)
+                            .build();
+                    RecruitmentBoardLike recruitmentLike = recruitmentBoardLikeRepository.findByRecruitmentBoardIdAndUserId(recruitmentBoardId, userId);
+                    recruitmentBoardLikeRepository.delete(recruitmentLike);
+                    recruitmentBoardRepository.save(updatedRecruitmentBoard);
+                    return null;
+                } else {
+                    throw new CommonException(ErrorCode.NOT_FOUND_LIKE);
+                }
             }
         }
         return null;
